refactor whole calc code to use functions and enum logic with type/ err handling + make f64

+++ crates/maths/src/logic/simple.rs
@@ -1,15 +1,15 @@
-pub fn add(a: i32, b: i32) -> i32 {
+pub fn add(a: f64, b: f64) -> f64 {
     a + b
 }
 
-pub fn subtract(a: i32, b: i32) -> i32 {
+pub fn subtract(a: f64, b: f64) -> f64 {
     a - b
 }
 
-pub fn multiply(a: i32, b: i32) -> i32 {
+pub fn multiply(a: f64, b: f64) -> f64 {
     a * b
 }
 
-pub fn divide(a: i32, b: i32) -> i32 {
+pub fn divide(a: f64, b: f64) -> f64 {
     a / b
 }
